+++ CreatedWidgetsSource/ol3-heatmap-widget/src/js/ol3-map-widget.js
@@ -10,15 +10,77 @@
         return url.href;
     };
 
-    var CORE_LAYERS = {
-        WIKIMEDIA: new ol.layer.Tile({
-            source: new ol.source.OSM({
-                url: "https://maps.wikimedia.org/osm-intl/{z}/{x}/{y}.png"
-            })
-        }),
-        OSM: new ol.layer.Tile({
-            source: new ol.source.OSM()
-        }),
+    var layerInfo = function layerInfo(method, url, done) {
+        var xhr = new XMLHttpRequest();
+        xhr.open(method, url);
+        xhr.onload = function () {
+            done(null, xhr.response);
+        };
+        xhr.onerror = function () {
+            done(xhr.response);
+        };
+        xhr.send();
+    };
+
+    var createLayers = function createLayers(result, layerArray) {
+        var BASE_LAYERS = {};
+
+        layerArray.forEach(function (layer) {
+            var identifier = layer.Identifier;
+            var layerOptions = ol.source.WMTS.optionsFromCapabilities(result, {
+                layer: identifier,
+                matrixSet: 'google3857',
+                requestEncoding: 'REST',
+                style: 'normal'
+            });
+            BASE_LAYERS[identifier] = new ol.layer.Tile({
+                source: new ol.source.WMTS(layerOptions)
+            });
+        });
+        return BASE_LAYERS;
+    };
+
+    var createLayerOptions = function createLayerOptions(layerArray, map, baseLayers) {
+        var button = document.getElementById('button');
+        var rect = button.getBoundingClientRect();
+        var items = [];
+        var refpos = {
+            top: rect.bottom,
+            left: rect.left,
+            width: 0,
+            height: 0
+        };
+
+        layerArray.forEach(function (layer) {
+            var identifier = layer.Identifier;
+            var title = layer.Title;
+
+            var item = new StyledElements.MenuItem(title, function () {
+                map.getLayers().removeAt(0);
+                map.getLayers().insertAt(0, baseLayers[identifier]);
+            });
+            items.push(item);
+        });
+
+        var popover = new StyledElements.PopupMenu();
+        for (var i = 0; i < items.length; i++) {
+            popover.append(items[i]);
+        }
+
+        button.addEventListener('click', function () {
+            popover.show(refpos);
+        });
+    };
+
+    var getLayerByName = function getLayerByName(name, map) {
+        var resultLayer = null;
+        map.getLayers().forEach(function (layer) {
+            var layerName = layer.get('name');
+            if (layerName === name) {
+                resultLayer = layer;
+            }
+        });
+        return resultLayer;
     };
 
     var build_basic_style = function build_basic_style(options) {
@@ -60,17 +122,45 @@
         });
     };
 
+    var createEmptyLayer = function createEmptyLayer(layerName) {
+        var layerSource = new ol.source.Vector({});
+        var layer = new ol.layer.Vector({source: layerSource, style: DEFAULT_MARKER});
+        layer.set('name', layerName);
+        return layer;
+    };
+
+    var createEmptyHeatmap = function createEmptyHeatmap(heatmapName) {
+        var mapSource = new ol.source.Vector();
+        var heatmapLayer = new ol.layer.Heatmap({
+            source: mapSource,
+            radius: 25,
+            blur: 20
+        });
+        heatmapLayer.set('name', heatmapName);
+        return heatmapLayer;
+    };
+
+    var toggleLayer = function toggleLayer(checkboxId, layerName, map) {
+        var checkbox = document.getElementById(checkboxId);
+        checkbox.addEventListener('change', function () {
+            var layer = getLayerByName(layerName, map);
+            if (this.checked) {
+                layer.setVisible(true);
+            } else {
+                layer.setVisible(false);
+            }
+        });
+    };
+
     // Create the default Marker style
     var DEFAULT_MARKER = build_basic_style();
 
     var Widget = function Widget() {
-        this.layers_widget = null;
         this.base_layer = null;
         this.layers = {};
     };
 
     Widget.prototype.init = function init() {
-        var initialLayer = CORE_LAYERS.WIKIMEDIA;
         var initialCenter = MashupPlatform.prefs.get("initialCenter").split(",").map(Number);
         if (initialCenter.length != 2 || !Number.isFinite(initialCenter[0]) || !Number.isFinite(initialCenter[0])) {
             initialCenter = [0, 0];
@@ -78,24 +168,87 @@
 
         this.map = new ol.Map({
             target: document.getElementById('map'),
-            layers: [
-                initialLayer
-            ],
             view: new ol.View({
                 center: ol.proj.transform(initialCenter, 'EPSG:4326', 'EPSG:3857'),
                 zoom: parseInt(MashupPlatform.prefs.get('initialZoom'), 10)
             })
         });
+
+        var map = this.map;
+
+        layerInfo('GET', 'https://www.basemap.at/wmts/1.0.0/WMTSCapabilities.xml', function (err, res) {
+            var result = new ol.format.WMTSCapabilities().read(res);
+            var resultContents = result.Contents;
+            var layerArray = resultContents.Layer;
+            var baseLayers = createLayers(result, layerArray);
+
+            map.getLayers().insertAt(0, baseLayers["geolandbasemap"]);
+            createLayerOptions(layerArray, map, baseLayers);
+
+            if (err) {
+                MashupPlatform.widget.log(err);
+            }
+        });
+
+        var heatmapLayer = createEmptyHeatmap('heatmap');
+        var staticLayer = createEmptyLayer('static');
+        var mobileLayer = createEmptyLayer('mobile');
+
+        this.map.addLayer(heatmapLayer);
+        this.map.addLayer(staticLayer);
+        this.map.addLayer(mobileLayer);
+
+        toggleLayer('heatmap', 'heatmap', this.map);
+        toggleLayer('static', 'static', this.map);
+        toggleLayer('mobile', 'mobile', this.map);
+
+        // display popup on click
+        this.map.on('click', function (event) {
+            var feature = this.map.forEachFeatureAtPixel(event.pixel,
+                function (feature, layer) {
+                    if (layer.get('name') != 'heatmap') {
+                        return feature;
+                    }
+                });
+
+            if (feature != null && feature !== this.selected_feature) {
+                this.select_feature(feature);
+            } else if (feature !== this.selected_feature) {
+                if (this.popover != null) {
+                    this.popover.hide();
+                    this.popover = null;
+                }
+            }
+        }.bind(this));
+
+        // change mouse cursor when over marker
+        this.map.on('pointermove', function (event) {
+            if (event.dragging) {
+                if (this.popover != null) {
+                    this.popover.hide();
+                    this.popover = null;
+                }
+                return;
+            }
+            var pixel = this.map.getEventPixel(event.originalEvent);
+            var hit = this.map.hasFeatureAtPixel(pixel);
+            this.map.getTarget().style.cursor = hit ? 'pointer' : '';
+        }.bind(this));
     };
 
     Widget.prototype.registerPoI = function registerPoI(poi_info) {
+        var poiData = poi_info.data;
+        var poiType = poiData.type;
+        MashupPlatform.widget.log(poiType, MashupPlatform.log.INFO);
+
+        var layer = getLayerByName(poiType, this.map);
         var iconFeature, style;
-        iconFeature = this.vector_source.getFeatureById(poi_info.id);
+        iconFeature = layer.getSource().getFeatureById(poi_info.id);
 
         if (iconFeature == null) {
             iconFeature = new ol.Feature();
             iconFeature.setId(poi_info.id);
-            this.vector_source.addFeature(iconFeature);
+            layer.getSource().addFeature(iconFeature);
         }
 
         iconFeature.set('data', poi_info.data);
@@ -145,74 +298,45 @@
         iconFeature.setStyle(style);
     };
 
-    Widget.prototype.addHeatmap = function addHeatmap(entities) {
-        var entitiesJson = JSON.parse(entities);
-        var pointFeature;
-        var mapData = new ol.source.Vector();
+    Widget.prototype.addHeatmap = function addHeatmap(poi_info) {
+        var heatmapLayer = getLayerByName('heatmap', this.map);
+        var heatmapFeature;
+        heatmapFeature = heatmapLayer.getSource().getFeatureById(poi_info.id);
 
-        entitiesJson.forEach(function (entity) {
-            MashupPlatform.widget.log(entity, MashupPlatform.log.INFO);
-            var temperature = entity["temperature"];
-            var locationParts = entity["location"].split(new RegExp(',\\s*'));
-            var long = parseFloat(locationParts[0]);
-            var lat = parseFloat(locationParts[1]);
-
-            pointFeature = new ol.Feature({
-                geometry: new ol.geom.Point(ol.proj.transform([lat, long], 'EPSG:4326',
-                    'EPSG:3857'))
-            });
-
-            if (temperature <= 0) {
-                pointFeature.set('weight', 0.2);
-            } else if (temperature > 0 && temperature <= 10) {
-                pointFeature.set('weight', 0.4);
-            } else if (temperature > 10 && temperature <= 20) {
-                pointFeature.set('weight', 0.6)
-            } else if (temperature > 20 && temperature <= 30) {
-                pointFeature.set('weight', 0.8)
-            } else {
-                pointFeature.set('weight', 1)
-            }
-
-            mapData.addFeature(pointFeature);
-        });
-
-        var heatmapLayer = new ol.layer.Heatmap({
-            source: mapData,
-            radius: 25,
-            blur: 20
-        });
-        heatmapLayer.set('name', 'heatmap');
-        this.map.addLayer(heatmapLayer);
-    };
-
-    Widget.prototype.addLayer = function addLayer(layer_info) {
-        var layer, params, service_url;
-
-        params = {
-            'LAYERS': layer_info.name,
-            'VERSION': layer_info.version
-        };
-
-        service_url = new URL(layer_info.url);
-        if (document.location.protocol === 'https:' && service_url.protocol !== 'https:') {
-            service_url = MashupPlatform.http.buildProxyURL(service_url.href);
-        } else {
-            service_url = layer_info.url;
+        if (heatmapFeature == null) {
+            heatmapFeature = new ol.Feature();
+            heatmapFeature.setId(poi_info.id);
+            heatmapLayer.getSource().addFeature(heatmapFeature);
         }
 
-        layer = new ol.layer.Image({
-            extent: layer_info.extent,
-            crossOrigin: 'anonymous',
-            source: new ol.source.ImageWMS({
-                url: service_url,
-                params: params,
-                projection: layer_info.projection
-            })
-        });
-        this.map.addLayer(layer);
+        heatmapFeature.set('data', poi_info.data);
+        heatmapFeature.set('title', poi_info.title);
+        heatmapFeature.set('content', poi_info.infoWindow);
+        if ('location' in poi_info) {
+            var geometry = this.geojsonparser.readGeometry(poi_info.location).transform('EPSG:4326', 'EPSG:3857');
+            heatmapFeature.setGeometry(new ol.geom.Point(geometry.getExtent()));
+        } else {
+            heatmapFeature.setGeometry(
+                new ol.geom.Point(
+                    ol.proj.transform([poi_info.currentLocation.lng, poi_info.currentLocation.lat], 'EPSG:4326', 'EPSG:3857')
+                )
+            );
+        }
 
-        this.layers[layer_info.url + '#' + layer_info.name] = layer;
+        var mapData = poi_info.data;
+        var temperature = mapData['temperature'];
+
+        if (temperature <= 0) {
+            heatmapFeature.set('weight', 0.2);
+        } else if (temperature > 0 && temperature <= 10) {
+            heatmapFeature.set('weight', 0.4);
+        } else if (temperature > 10 && temperature <= 20) {
+            heatmapFeature.set('weight', 0.6)
+        } else if (temperature > 20 && temperature <= 30) {
+            heatmapFeature.set('weight', 0.8)
+        } else {
+            heatmapFeature.set('weight', 1)
+        }
     };
 
     Widget.prototype.center_popup_menu = function center_popup_menu(feature) {